client uses sdk resource handler
